Use a proper long type for cache byte size, but you'll need to tweak COConfigurationManager.CONFIG_CACHE_SIZE_MAX_MB limit check if you want to go beyond 1GB.
git-svn-id: http://svn.vuze.com/public/client@24691 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/diskmanager/cache/impl/CacheFileManagerImpl.java
@@ -156,7 +156,7 @@ CacheFileManagerImpl
 		
 			// units are MB
 		
-		int		size			= 1024*1024*COConfigurationManager.getIntParameter( "diskmanager.perf.cache.size" );
+		long		size			= 1024L * 1024L * COConfigurationManager.getIntParameter( "diskmanager.perf.cache.size" );
 		
 			// units are KB
